+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## [0.2.0]
+
+ * [fix] Error handling
+ * [fix] Terminating of queries on stream drop
+ * [new] Sink and source connectors
 ## [0.1.1]
 
  * [fix] KSQL Statements were not executed